Move graphs to separate JS file
