+++ app/javascript/packs/pages/binx_admin.js
@@ -2,6 +2,7 @@ import log from "../utils/log";
 import moment from "moment-timezone";
 import LoadFancySelects from "../utils/LoadFancySelects";
 import BinxAdminInvoices from "./binx_admin_invoices.js";
+import BinxGraphs from "./binx_graphs.js"
 
 function BinxAdmin() {
   return {
@@ -16,8 +17,8 @@ function BinxAdmin() {
           .css("max-width", "100%");
       }
       if ($(".calendar-box")[0]) {
-        this.setCustomGraphStartAndSlide();
-        this.changeGraphCalendarBox();
+        const binxGraphs = BinxGraphs()
+        binxGraphs.init()
       }
       if ($("#use_image_for_display").length > 0) {
         this.useBikeImageForDisplay();
@@ -34,33 +35,6 @@ function BinxAdmin() {
       LoadFancySelects();
 
       this.enablePeriodSelection();
-    },
-
-    changeGraphCalendarBox() {
-      $("select#graph_date_option").on("change", e => {
-        e.preventDefault();
-        this.setCustomGraphStartAndSlide();
-      });
-    },
-
-    startGraphTimeSet() {
-      const graphSelected = $("select#graph_date_option")[0].value.split(",");
-      const amount = Number(graphSelected[0]);
-      const unit = graphSelected[1];
-      $("#start_at").val(
-        moment()
-          .subtract(amount, unit)
-          .format("YYYY-MM-DDTHH:mm")
-      );
-    },
-
-    setCustomGraphStartAndSlide() {
-      if ($("select#graph_date_option")[0].value === "custom") {
-        $(".calendar-box").slideDown();
-      } else {
-        $(".calendar-box").slideUp();
-        this.startGraphTimeSet();
-      }
     },
 
     useBikeImageForDisplay() {
